Moved inheritance tests to own class
